% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrot_functions_rsq.R
\name{AUC}
\alias{AUC}
\title{Area Under the Curve}
\usage{
AUC(probs, class)
}
\arguments{
\item{probs}{probabilities}

\item{class}{outcomes}
}
\value{
A value for AUC
}
\description{
Function enables efficient computation of area under receiver operating curve (AUC). Source: \url{https://stat.ethz.ch/pipermail/r-help/2005-September/079872.html}
}
\examples{
AUC(runif(100,0,1),rbinom(100,1,0.3))
}
