% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrot_functions.R
\name{comb}
\alias{comb}
\title{Combining in a list}
\usage{
comb(...)
}
\arguments{
\item{...}{an argument of \code{mapply} used by this function}
}
\description{
Function for combining outputs in a list
}
\examples{
#array of numbers to be separated in a list

a<-1:4

#running the function

comb(a)
}
\seealso{
Function \code{\link[base]{mapply}}
}
