% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrot_functions.R
\name{cub}
\alias{cub}
\title{Three-way interactions and squares}
\usage{
cub(A, n = 1000)
}
\arguments{
\item{A}{set of predictors}

\item{n}{first \code{n} predictors, whose interactions with the rest should be taken into account, defaults to all of the predictors}
}
\value{
Returns the predictors including their squares, pairwise interactions, cubes and three-way interactions
}
\description{
Function transforms a set of predictors into a set of predictors, their squares, pairwise interactions, cubes and three-way interactions
}
\examples{
cub(cbind(1:100,rnorm(100),runif(100),rnorm(100,0,2)))


}
