% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrot_functions_rsq.R
\name{find_int}
\alias{find_int}
\title{Finding the interacting terms based on the index}
\usage{
find_int(ind,N)
}
\arguments{
\item{ind}{index to transform}

\item{N}{number of interacting variables}
}
\value{
Returns two or three indices corredsponding to a combination of variables written under the given index
}
\description{
Function transforms an index of an array of two- or three-way interactions into two or three indices corresponding to the interacting variables
}
\examples{
find_int(28,9)
}
