% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carrot_functions_rsq.R
\name{quadr}
\alias{quadr}
\title{Pairwise interactions and squares}
\usage{
quadr(A, n = 1000)
}
\arguments{
\item{A}{set of predictors}

\item{n}{first \code{n} predictors, whose interactions with the rest should be taken into account, defaults to all of the predictors}
}
\value{
Returns the predictors including their squares and pairwise interactions
}
\description{
Function transforms a set of predictors into a set of predictors, their squares and pairwise interactions
}
\examples{
quadr(cbind(1:100,rnorm(100),runif(100),rnorm(100,0,2)))
}
